added dependency injection to FormController
