+++ app/Http/Controllers/FormController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use App\Services\DogDataService;
+use App\Services\ExternalApiService;
 use Illuminate\Http\Request;
 use App\Selection;
 use App\User;
@@ -9,14 +11,21 @@ use App\Http\Controllers\BreedController;
 
 class FormController extends Controller
 {
+    protected $externalApiService;
+    protected $dogDataService;
+    public function __construct(ExternalApiService $externalApiService, DogDataService $dogDataService)
+    {
+        $this->externalApiService = $externalApiService;
+        $this->dogDataService = $dogDataService;
+    }
+
     public function saveUserRecordToEmail($email='etha2@gmail.com', $miles=44, $breed='Bearded Collie', $zip = "91306"){
-        $breedController = new BreedController();
-        $breedArray = $breedController->getExternalDataForBreed($zip,$breed);
+        $breedArray = $externalApiService->getExternalDataForBreed($zip,$breed);
         $selection =
             Selection::create([
-                'breed_id' => $breedController->getBreedId($breed),
+                'breed_id' => $dogDataService->getBreedId($breed),
                 'zip' => $zip,
-                'highest_breed_id' => $breedController->getLargestBreedId($breedArray),
+                'highest_breed_id' => $dogDataService->getLargestBreedId($breedArray),
                 'max_miles' => $miles,
                 'match'     => false
             ])
